Final version uploaded to AFS 11-30-18
I used your function for the mean, however, your function for the standard deviation returned a value for each datapoint, not a value for each feature and returned an index error.  Since the program takes so long to run, I also uploaded the output in a text file.  Both files were added to my AFS this morning.